requirements.txt
click==7.1.2
sqlparams
Flask==1.1.2
Jinja2==2.11.2
MarkupSafe==1.1.1
Werkzeug==1.0.1
gunicorn
python-dotenv
sqlalchemy